requirements.txt
# ===========================
# requirements.txt
# ===========================
# Core libraries for stateful, multi-agent, and LLM applications
langgraph==1.0.2                   # For building stateful, multi-agent workflows
langchain==1.0.3                  # Core LangChain abstractions
langchain-core==1.0.3             # LangChain core utilities
langchain-community==0.4.1        # Community-contributed LangChain integrations
langchain-huggingface==1.0.1       # HuggingFace integration for LangChain
langchain-google-genai==3.0.1     # Google Gemini models integration for LangChain

# Environment and configuration management
pyyaml                     # For YAML configuration files
python-dotenv              # For loading environment variables from .env

# Retrieval-Augmented Generation (RAG) dependencies
requests                   # HTTP requests (e.g., PubMed API)
datasets                   # Downloading datasets (e.g., PubMedQA from Hugging Face)
faiss-cpu                  # In-memory vector store for RAG
tiktoken                   # Tokenizer for text splitting

# Search and embedding tools
duckduckgo-search          # DuckDuckGo web search integration
sentence-transformers      # Sentence embeddings (HuggingFace models)
ddgs                       # Alternative DuckDuckGo search API
beautifulsoup4             # XML/HTML parsing
lxml                       # XML parser for BeautifulSoup

# App and UI
streamlit                  # Web app framework for Python

# ===========================
# End of requirements.txt
# ===========================